requirements.txt

streamlit
moviepy
pillow
requests
numpy
